pipebot/requirements.txt
boto3>=1.34.0
colored>=2.2.0
dataclasses>=0.6
typing-extensions>=4.9.0
chromadb==0.6.3
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyJWT==2.8.0
azure-identity==1.15.0
tiktoken==0.5.2
bs4==0.0.2
redis==5.2.1

# Scientific computing dependencies
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
scipy==1.13.1
scikit-learn==1.6.1
sympy==1.13.1
mpmath==1.3.0
pymongo==4.10.1
python-dateutil==2.8.2 